Refactored controllers to use Repositories, and added User and Product repositories.

+++ web-portal-api/webportalapp.domain/Entities/User.cs
@@ -2,7 +2,7 @@
 {
     public class User
     {
-        public long UID { get; set; }   
+        public int UID { get; set; }   
         public string Username { get; set; }
         public string Password { get; set; }
         public string Email { get; set; }
